Create and apply `OpTable` class for runtimes
