+++ merlin/systems/dag/runtimes/__init__.py
@@ -13,6 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# flake8: noqa
 
+# flake8: noqa
 from .base_runtime import Runtime
