+++ merlin/systems/dag/runtimes/triton/runtime.py
@@ -33,6 +33,7 @@ from merlin.systems.dag.ops.compat import (
 )
 from merlin.systems.dag.ops.workflow import TransformWorkflow
 from merlin.systems.dag.runtimes import Runtime
+from merlin.systems.dag.runtimes.op_table import OpTable
 from merlin.systems.dag.runtimes.triton.ops.operator import TritonOperator, add_model_param
 from merlin.systems.dag.runtimes.triton.ops.workflow import TransformWorkflowTriton
 
@@ -53,26 +54,26 @@ except ImportError:
     ...
 
 
-TRITON_OP_TABLE = {}
-TRITON_OP_TABLE[TransformWorkflow] = TransformWorkflowTriton
+TRITON_OP_TABLE = OpTable()
+TRITON_OP_TABLE.register(TransformWorkflow, TransformWorkflowTriton)
 
 if cuml_ensemble or lightgbm or sklearn_ensemble or treelite_sklearn or xgboost:
     from merlin.systems.dag.ops.fil import PredictForest
     from merlin.systems.dag.runtimes.triton.ops.fil import PredictForestTriton
 
-    TRITON_OP_TABLE[PredictForest] = PredictForestTriton
+    TRITON_OP_TABLE.register(PredictForest, PredictForestTriton)
 
 if tensorflow:
     from merlin.systems.dag.ops.tensorflow import PredictTensorflow
     from merlin.systems.dag.runtimes.triton.ops.tensorflow import PredictTensorflowTriton
 
-    TRITON_OP_TABLE[PredictTensorflow] = PredictTensorflowTriton
+    TRITON_OP_TABLE.register(PredictTensorflow, PredictTensorflowTriton)
 
 if torch:
     from merlin.systems.dag.ops.pytorch import PredictPyTorch
     from merlin.systems.dag.runtimes.triton.ops.pytorch import PredictPyTorchTriton
 
-    TRITON_OP_TABLE[PredictPyTorch] = PredictPyTorchTriton
+    TRITON_OP_TABLE.register(PredictPyTorch, PredictPyTorchTriton)
 
 
 class TritonExecutorRuntime(Runtime):
